fix more diff

+++ test_infer.py
@@ -45,7 +45,7 @@ if __name__ == '__main__':
         model_path,
         torch_dtype=torch.bfloat16,
         attn_implementation="flash_attention_2",
-    ).to("cuda")
+    ).to("cuda", dtype=torch.bfloat16)
 
     vision_path = "/mllm_native/sunyunxiao.syx/assets/"
